Do not assume `argon2` based hashes
The `extractArgon2ParametersFromEncodedPassword` assumes `argon2` based hashes and hence fails with an "Index out of bound" exception if the hash does not include a `$` sign. The introduced test checks this assumption to short cut testing the policy. If the stored hash is not an `argon2`-hash than the policy is violated and a rehash is required.